pyproject.toml
[project]
name = "presentation-manim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "manim>=0.19.0",
    "manim-slides>=5.5.0",
    "manimpango>=0.6.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pycairo",
]
